feat(hooks): 🍎 merge utility functions into hooks
